requirements.txt
langchain-community
langchain-core
langchain-experimental
langgraph
langchain-postgres
langchain-text-splitters
langsmith
ollama
openai
pdfplumber
pgvector
faiss-cpu
sentence-transformers
rank_bm25
gradio
notebook
utils
psutil